243: move JsonRquest to request and JsonReply to response
