+++ src/rpc_client/eth_rpc_error/mod.rs
@@ -1,6 +1,5 @@
 use crate::logs::DEBUG;
-use crate::rpc_client::eth_rpc::{JsonRpcReply, JsonRpcResult};
-use crate::rpc_client::json::responses::SendRawTransactionResult;
+use crate::rpc_client::json::responses::{JsonRpcReply, JsonRpcResult, SendRawTransactionResult};
 use ic_canister_log::log;
 use crate::rpc_client::json::Hash;
