+++ src/rpc_client/json/mod.rs
@@ -98,3 +98,4 @@ impl std::str::FromStr for Hash {
 }
 
 impl HttpResponsePayload for Hash {}
+
