+++ src/rpc_client/json/responses.rs
@@ -9,6 +9,7 @@ use candid::Deserialize;
 use ic_ethereum_types::Address;
 use serde::Serialize;
 use std::fmt::{Display, Formatter};
+use evm_rpc_types::{JsonRpcError, RpcError};
 use crate::rpc_client::json::{FixedSizeData, Hash};
 
 #[derive(Debug, Clone, Serialize, Deserialize, Eq, PartialEq)]
@@ -253,5 +254,42 @@ pub struct Data(#[serde(with = "ic_ethereum_types::serde_data")] pub Vec<u8>);
 impl AsRef<[u8]> for Data {
     fn as_ref(&self) -> &[u8] {
         &self.0
+    }
+}
+
+#[derive(Debug, PartialEq, Eq, Serialize, Deserialize)]
+pub struct JsonRpcReply<T> {
+    pub id: u64,
+    pub jsonrpc: String,
+    #[serde(flatten)]
+    pub result: JsonRpcResult<T>,
+}
+
+/// An envelope for all JSON-RPC replies.
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub enum JsonRpcResult<T> {
+    #[serde(rename = "result")]
+    Result(T),
+    #[serde(rename = "error")]
+    Error { code: i64, message: String },
+}
+
+impl<T> JsonRpcResult<T> {
+    pub fn unwrap(self) -> T {
+        match self {
+            Self::Result(t) => t,
+            Self::Error { code, message } => panic!(
+                "expected JSON RPC call to succeed, got an error: error_code = {code}, message = {message}"
+            ),
+        }
+    }
+}
+
+impl<T> From<JsonRpcResult<T>> for Result<T, RpcError> {
+    fn from(result: JsonRpcResult<T>) -> Self {
+        match result {
+            JsonRpcResult::Result(r) => Ok(r),
+            JsonRpcResult::Error { code, message } => Err(JsonRpcError { code, message }.into()),
+        }
     }
 }
